mockup AsyncActionPlugin derive macro
