+++ packages/task/src/lib.rs
@@ -1,7 +1,8 @@
 mod async_action;
 
 pub use crate::async_action::AsyncActionPool;
+pub use dip_macro::AsyncActionPlugin;
 
 pub mod prelude {
-    pub use crate::AsyncActionPool;
+    pub use crate::{AsyncActionPlugin, AsyncActionPool};
 }
